feat: Maya integration for windows (#167)
* Can install plugin for Maya on Windows
* Save plugin installation info in local AppData folder
* Fix path-related functions on Windows
* Handle errors where we are not allowed to write installation info
* Add IconPath to DCCToolInfo, which is filled together with AppPath
* Shows DCC tool images in MeshSync project settings. Only works for png files for now.

Others:
* Update the ProjectSettings of the test project
* A bit of preparation for 3DS Max integration: finding config folder, etc

+++ Editor/Scripts/DCCIntegration/MayaIntegrator.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 namespace UnityEditor.MeshSync {
 
 internal class MayaIntegrator : BaseDCCIntegrator {
-
+    
     internal MayaIntegrator(DCCToolInfo dccToolInfo) : base(dccToolInfo) { }
 
 //----------------------------------------------------------------------------------------------------------------------
@@ -17,72 +17,89 @@ internal class MayaIntegrator : BaseDCCIntegrator {
     }
 
 //----------------------------------------------------------------------------------------------------------------------
-    protected override DCCPluginInstallInfo ConfigureDCCTool(DCCToolInfo dccToolInfo, string configFolder, 
-        string localPluginPath) 
+    protected override bool ConfigureDCCTool(DCCToolInfo dccToolInfo, string configFolder, string localPluginPath) 
     {
-
         string tempPath = FileUtil.GetUniqueTempPathInProject();
+        
         Directory.CreateDirectory(tempPath);
         ZipUtility.UncompressFromZip(localPluginPath, null, tempPath);
 
         string srcRoot = Path.Combine(tempPath, Path.GetFileNameWithoutExtension(localPluginPath));
         if (!Directory.Exists(srcRoot)) {
-            Debug.LogError("[MeshSync] Failed to install DCC Plugin for Maya");
-            return null;
+            return false;
         }
 
         const string AUTOLOAD_SETUP = "pluginInfo -edit -autoload true MeshSyncClientMaya;";
         const string SHELF_SETUP = "UnityMeshSync_Shelf;";
         const string MAYA_CLOSE_COMMAND = "scriptJob -idleEvent quit;";
         const string FINALIZE_SETUP = AUTOLOAD_SETUP + SHELF_SETUP + MAYA_CLOSE_COMMAND;
-
-        int exitCode = 0;
-
+        
+        string copySrcFolder  = srcRoot;
+        string copyDestFolder = configFolder;
+        string argFormat = null;
+        string loadPluginCmd = null;
+            
+            
         switch (Application.platform) {
             case RuntimePlatform.WindowsEditor: {
-                // If MAYA_APP_DIR environment variable is setup, copy the modules directory there.
-                //     If not, go to %USERPROFILE%\Documents\maya in Windows Explorer, and copy the modules directory there.
-                //string commandString = "-command \"{0}\"";
+                //C:\Users\Unity\Documents\maya\modules
+                const string FOLDER_PREFIX = "modules";
+                copySrcFolder  = Path.Combine(srcRoot, FOLDER_PREFIX);
+                copyDestFolder = Path.Combine(configFolder, FOLDER_PREFIX);                
+                
+                argFormat = "-command \"{0}\"";
+
+                //Maya script only supports '/' as PathSeparator
+                //Example: loadPlugin """C:/Users/Unity/Documents/maya/modules/UnityMeshSync/2019/plug-ins/MeshSyncClientMaya.mll""";
+                string mayaPluginPath = Path.Combine(copyDestFolder, "UnityMeshSync", dccToolInfo.DCCToolVersion, 
+                    @"plug-ins\MeshSyncClientMaya.mll").Replace('\\','/');
+                loadPluginCmd = "loadPlugin \"\"\"" + mayaPluginPath + "\"\"\";";
                 break;
             }
             case RuntimePlatform.OSXEditor: {
                 
-                //Copy files
-                const string MOD_FILE = "UnityMeshSync.mod";
-                string SCRIPT_FOLDER = Path.Combine("UnityMeshSync",dccToolInfo.DCCToolVersion);
-                File.Copy(Path.Combine(srcRoot,MOD_FILE), Path.Combine(configFolder,MOD_FILE), true);
-                FileUtility.CopyRecursive(Path.Combine(srcRoot, SCRIPT_FOLDER), 
-                                          Path.Combine(configFolder, SCRIPT_FOLDER),
-                                          true);
-                
                 //Setup Auto Load
-                string commandString = @"-command '{0}'";
+                argFormat = @"-command '{0}'";
                 //Example: "/Users/Shared/Autodesk/Modules/maya/UnityMeshSync/2020/plug-ins/MeshSyncClientMaya.bundle";
-                string loadPlugin = "loadPlugin \"" + configFolder + "/UnityMeshSync/"+ dccToolInfo.DCCToolVersion 
-                                    + "/plug-ins/MeshSyncClientMaya.bundle\";";
-                
-                string argument = string.Format(commandString, loadPlugin+FINALIZE_SETUP);
-                exitCode = SetupAutoLoadPlugin(dccToolInfo.AppPath, argument);
+                loadPluginCmd = "loadPlugin \"" + configFolder + "/UnityMeshSync/"+ dccToolInfo.DCCToolVersion 
+                                       + "/plug-ins/MeshSyncClientMaya.bundle\";";
                 
                 break;
             }
             case RuntimePlatform.LinuxEditor: {
                 // Copy the modules directory to ~/maya/<maya_version)            
-                break;
+                throw new NotImplementedException();
             }
             default: {
                 throw new NotImplementedException();
             }
         }
-        
+
+        //Copy files to config folder
+        const string MOD_FILE = "UnityMeshSync.mod";
+        string scriptFolder = Path.Combine("UnityMeshSync",dccToolInfo.DCCToolVersion);
+        string srcModFile = Path.Combine(copySrcFolder, MOD_FILE);
+        if (!File.Exists(srcModFile)) {
+            return false;
+        }
+        try {
+            Directory.CreateDirectory(copyDestFolder);
+            File.Copy(srcModFile, Path.Combine(copyDestFolder, MOD_FILE), true);
+            FileUtility.CopyRecursive(Path.Combine(copySrcFolder, scriptFolder),
+                Path.Combine(copyDestFolder, scriptFolder),
+                true);
+        } catch {
+            return false;
+        }
+
+        //Auto Load
+        string arg = string.Format(argFormat, loadPluginCmd+FINALIZE_SETUP);
+        int exitCode = SetupAutoLoadPlugin(dccToolInfo.AppPath, arg);
 
         //Cleanup
         FileUtility.DeleteFilesAndFolders(tempPath);
 
-        if (0 != exitCode)
-            return null;
-        
-        return new DCCPluginInstallInfo(dccToolInfo.DCCToolVersion);
+        return (0 == exitCode);
     }
     
 //----------------------------------------------------------------------------------------------------------------------    
@@ -90,10 +107,21 @@ internal class MayaIntegrator : BaseDCCIntegrator {
     protected override string FindConfigFolder() {
         switch (Application.platform) {
             case RuntimePlatform.WindowsEditor: {
-                // If MAYA_APP_DIR environment variable is setup, copy the modules directory there.
-                //     If not, go to %USERPROFILE%\Documents\maya in Windows Explorer, and copy the modules directory there.
-                break;
+
+                //If MAYA_APP_DIR environment variable is setup, use that config folder
+                //If not, use %USERPROFILE%\Documents\maya 
+                string path = Environment.GetEnvironmentVariable("MAYA_APP_DIR");
+                if (!string.IsNullOrEmpty(path))
+                    return path;
+
+                path = Directory.GetParent(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)).FullName;
+                if (Environment.OSVersion.Version.Major >= 6) {
+                    path = Directory.GetParent(path).ToString();
+                }
+                path+= @"\Documents\maya";
+                return path;
             }
+
             case RuntimePlatform.OSXEditor: { return "/Users/Shared/Autodesk/modules/maya"; }
             case RuntimePlatform.LinuxEditor: {
                 // Copy the modules directory to ~/maya/<maya_version)            
@@ -113,7 +141,7 @@ internal class MayaIntegrator : BaseDCCIntegrator {
     // [SecurityPermission(SecurityAction.LinkDemand, Flags = SecurityPermissionFlag.UnmanagedCode)]
     int SetupAutoLoadPlugin(string mayaPath, string startArgument) {
         int exitCode = 0;
-
+        
         try {
             if (!System.IO.File.Exists(mayaPath)) {
                 Debug.LogError("[MeshSync] No maya installation found at " + mayaPath);
